pyproject.toml
[tool.poetry]
name = "phishflood"
version = "0.1.0"
description = "Flood phishing websites with realistic fake credentials"
authors = ["Antonio Solana <solanav@proton.me>"]
license = "AGPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
playwright = "^1.40.0"
beautifulsoup4 = "^4.12.2"
zstandard = "^0.22.0"
pika = "^1.3.2"
django = "^5.0"
djangorestframework = "^3.14.0"
markdown = "^3.5.1"
django-filter = "^23.5"
pygments = "^2.17.2"
gunicorn = "^21.2.0"
psycopg2 = "^2.9.9"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
